+++ frontend/src/App.jsx
@@ -3,9 +3,9 @@ import Header from './components/Layout/Header'
 import NewTicketForm from './components/Tickets/NewTicketForm'
 import Tickets from './pages/Tickets'
 import FilterMenu from './pages/FilterMenu'
+import TicketDetails from './pages/TicketDetails'
 import Login from './pages/Login'
 import AuthGuard from './components/auth/AuthGuard'
-
 import './styles/globals.css';
 
 function App() {
@@ -24,6 +24,7 @@ function App() {
                 </>
               } />
               <Route path="/new-ticket" element={<NewTicketForm />} />
+              <Route path="/tickets/:id" element={<TicketDetails />} />
             </Routes>
           </main>
         </div>