+++ frontend/src/components/ui/Slider.jsx
@@ -10,7 +10,6 @@ const Slider = React.forwardRef(({ className, ...props }, ref) => {
     return 'bg-red-500';
   };
 
-
   return (
     <div className="relative w-full">
       <SliderPrimitive.Root
@@ -27,8 +26,6 @@ const Slider = React.forwardRef(({ className, ...props }, ref) => {
         </SliderPrimitive.Track>
         <SliderPrimitive.Thumb className="block h-5 w-5 rounded-full border-2 border-blue-900 bg-white ring-offset-background transition-transform duration-200 ease-in-out focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-blue-500 focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50" />
       </SliderPrimitive.Root>
-
-
     </div>
   );
 });